Completed creating seperate file for containing all the whole foods locations and making the mapContainer simply display that data to make adding multiple map types easier later on
